# src-tauri/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoopServe API",
    "description": "Looping media server API for listing assets, uploading media, and setting the live asset (Bitfocus Companion friendly).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "paths": {
    "/api/assets": {
      "get": {
        "tags": ["Assets"],
        "summary": "List assets",
        "description": "Returns all media assets and the current live asset id. Useful for Companion button labels.",
        "operationId": "listAssets",
        "responses": {
          "200": {
            "description": "Asset list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "liveId": {
                      "type": "string",
                      "nullable": true,
                      "example": "abc123xyz0"
                    },
                    "assets": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AssetWithLive" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Assets"],
        "summary": "Upload asset",
        "description": "Upload a video or image. Form field `file` is required; `name` is optional display name.",
        "operationId": "uploadAsset",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Video or image file"
                  },
                  "name": {
                    "type": "string",
                    "description": "Optional display name",
                    "example": "Worship Opener"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asset created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asset": { "$ref": "#/components/schemas/Asset" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid upload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/assets/{id}": {
      "put": {
        "tags": ["Assets"],
        "summary": "Update asset (keep id)",
        "description": "Replace the media file and/or rename the asset without changing its id. Useful for Companion buttons that target a stable asset id. Send multipart form-data with optional `file` and/or `name`, or JSON `{ \"name\": \"...\" }` for rename-only.",
        "operationId": "updateAsset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Replacement video or image"
                  },
                  "name": {
                    "type": "string",
                    "description": "Optional new display name"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Worship Opener v2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated asset (same id)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asset": { "$ref": "#/components/schemas/Asset" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid update",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Assets"],
        "summary": "Delete asset",
        "operationId": "deleteAsset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "deleted": { "$ref": "#/components/schemas/Asset" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Asset": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "abc123xyz0" },
          "name": { "type": "string", "example": "Worship Opener" },
          "type": { "type": "string", "enum": ["video", "image"] },
          "filename": { "type": "string" },
          "originalName": { "type": "string" },
          "url": { "type": "string", "example": "/media/1710000000-xyz.mp4" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AssetWithLive": {
        "allOf": [
          { "$ref": "#/components/schemas/Asset" },
          {
            "type": "object",
            "properties": {
              "live": { "type": "boolean" }
            }
          }
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}